Add files via upload
These changes  make the main screen visible and sized. It also creates a invisible box in the bottom left of the screen that is clickable.This area when clicked, opens us to a second screen. This will not be merged with the main branch until it has clearance by Elizabeth and Pamela. To clear this change, make a pull request and approve of the pull. Currently, I am just working to create a gui that can be used.